requirements.txt
py-clob-client==0.20.0
python-dotenv==0.19.2
pandas
gspread
gspread-dataframe
sortedcontainers
black==24.4.2
eth-account==0.13.0
eth-utils==4.1.1
poly_eip712_structs==0.0.1
py_order_utils==0.3.2
pytest==8.2.2
requests==2.32.3
websockets==12.0
cryptography==42.0.8
google-auth
web3==6.4.0